Translation fault in the zh-tw scopes.md.
In the "When not used in an assignment, return statement or as a function argument, the {...} notation will get interpreted as a block statement and not as an object literal" paragraph, the meaning in the English ver. is different in the Chinese ver.
I think there are translation faults which may lead to misunderstanding in the zh-tw version, so I make some corrections in that paragraph. Please check that. Thanks :-)